functions/requirements.txt
firebase-functions
firebase-admin
Pillow
numpy
opencv-python